refactor: move styles.tsx to styles/index.tsx
Other features use the `styles` directory for styles.
